[feat] Add post-collect
