+++ Dockerfile
@@ -6,7 +6,7 @@ ADD . /single_coordinator
 RUN . /usr/local/fix-venv-python3/bin/activate && pip install /single_coordinator && rm -rf /single_coordinator
 
 # Add shim and create symlink
-COPY collect_single_shim /usr/local/bin/collect_single_shim
+COPY dispatch_executable_shim.sh /usr/local/bin/dispatch_executable_shim
 RUN chmod 755 /usr/local/bin/collect_single_shim && ln -s /usr/local/bin/collect_single_shim /usr/bin/collect_single
 
-ENTRYPOINT ["/bin/dumb-init", "--", "/usr/local/sbin/bootstrap", "/usr/bin/collect_single"]
+ENTRYPOINT ["/bin/dumb-init", "--", "/usr/local/sbin/bootstrap", "/usr/bin/dispatch_executable_shim"]
